rpc: connection_creator: Fix formatting error
Warning currently prints out:

[WW 2017-05-03 15:15:30] [connection_creator.cpp:47@create_connection_for]
A second client tried to connect. Denied request as we already have oneand only allow a single client
                                                                       ^
Due to the way in which the code is formatted, we end up missing fudging
the words 'one' and 'and'.  Also the 3rd line in the source file seems
somewhat superfluous.  Let's fix that too.

Signed-off-by: Lee Jones <lee.jones@linaro.org>