android: service: platform_service: Read rotation angle from parcel
When passing window state, currently only the frame location and
stack & task IDs are contained in the parcel.  This patch also
receives the new 'rotation_angle' attribute.  This patch is critical
since it ensures each '4 Byte' value in this strongly ordered
sequence is extracted at the right time, in the correct way.

Signed-off-by: Lee Jones <lee.jones@linaro.org>