snap: make anbox work on systems with GLVND
Systems with GLVND load the actual EGL implementation in a different way
when the GL driver is dlopen'ed and therefor we need to tell GLVND which
vendor implementations are available on the system. As snapd is not yet
exposing this to the snap we have to ship a small set of configs on our
own until that is fixed.