lxc: bind relevant devices nodes into the container with shifted uid/gid
The simple bind mount we used before to forward necessary devices nodes
into the Android container missed shifting the uid/gid. Now we recreate
the devices nodes in a private directory and shift the uid/gid correctly
and then bind mount the resulting files into the container. This should
avoid problems we've have seen on some systems with not allowed access
to some of the device nodes.