Add AMD E-350 processor to the whitelist
Hi, I'd like to know if this is the right way to override the "It is missing support for the following features: SSE 4.1, SSE 4.2, SSSE 3" error message, I don't really know the language but after looking at other commits I decided to give it a try. Can somebody please let me know If I did it right? 

Thank you:)