session: let healthd provide a fake battery
Instruct healthd to fake battery level as it may take it from other connected
devices like mouse or keyboard and will incorrectly show a system popup to
shutdown the Android system because of low battery. This prevents any kind of
input as focus is bound to the system popup exclusively.

See https://github.com/anbox/anbox/issues/780 for further details.