container: fix container gateway and address prefix length
Use the custom gateway if present and extend the configuration option
container.network.address to allow both an IPv4 address and a prefix
length, or only an IPv4 address in which case the prefix length is set
to 24. Which means the following configuration options give the same
result.

snap set anbox container.network.address=192.168.250.2/24
snap set anbox container.network.address=192.168.250.2

Closes #1052