input: Ensure sockets are accessible from within the container
The input sockets are created within the user tmpfs (e.g. /run/user/...).
Files/sockets created there might not be accessible to other users.
Since the input sockets are bind-mounted into the Android container,
it will crash with "Permission denied" when it is unable to access them.

Make sure to give the created directory+socket files appropriate
permissions so they can be accessed within the Android container.

Similar code exists for the anbox_bridge and anbox_audio socket in
  - src/anbox/container/service.cpp (anbox_bridge)
  - src/anbox/audio/server.cpp (anbox_audio)