snap: fix failure on cloning SwiftShader repository
snapcraft uses '--recursive-submodules' option by default to clone all
dependencies if source-type is 'git'. Therefore the option leads the failure
on cloning the project by 'git-hooks' submodule. In the case, the failure is
asking username/password to do so.

Use 'override-pull' to prevent the issue as the following
1. clone the project w/o submodules.
2. use 'sed' command to remove 'git-hooks' submodule.
3. update submodules

'git-hooks' submodule is not necessary to build the project[1]

[1]: https://issuetracker.google.com/issues/190626829

Fix: https://github.com/anbox/anbox/issues/1922